Rock-Paper-Scissor Game
Rock, Paper, Scissors is a classic hand game often used as a simple decision-making tool. It involves two players simultaneously forming one of three shapes with their hands. The three shapes are:

Rock – a fist
Paper – an open hand
Scissors – a fist with the index and middle fingers extended, forming a V.

It is made by using simple HTML, CSS and JavaScript.